fix: rate exchange fix for Insufficient Liquidity
While determining the exchange rate, when a temporary error (usually T04 Insufficient Liquidity) is encountered, the connection should retry 2/3 the smallest amount sent that error'd, _not_ the smallest amount that was sent in the entire volley.

For example, if the first volley is `[1, 1000, 1000000, 1000000000, 1000000000000]` and all but the `1` return `T04`, it shouldn't retry 1*(2/3), it should retry 1000*(2/3).