fix: clean up the determineExchangeRate loop
Given the following scenario:

- Client connects to server.
- Client negotiates exchange rate with server.
- Server tries to negotiate an exchange rate with the client. But it doesn't work due to `Txx` errors.
- The client sends money, then closes the connection.
- The server closes the connection.

`determineExchangeRate()` keeps going. Because of the backoff settings and many attempts, it will continue for ~7.4 minutes. Eventually it runs out of attempts and throws the `Unable to establish connection, no packets meeting …` error.

```javascript
// Cumulative delay.
let total = 0
// Initial delay.
let step = 100

for (let i = 1; i < 20; i++) {
  total += step
  step *= 1.5
}

console.log(total / 1000 / 60)
// > 7.386116666666666 minutes
```

To fix: When the connection ends, if a `determineExchangeRate` is running, abort it immediately.