fix(connection): log uncaught rejection on end
Fixes a bug that Uphold ran into. An uncaught rejection was bubbling up from `end()`.

I was unable to reproduce locally or as a test case. My best guess is that the error is triggered by a concurrent `startSendLoop` (i.e. the exchange rate probe) that triggers an error event, while simultaneously a `ConnectionClose` frame is being handled. The `error` is caught by the `.once('error', reject)` in the `end` method, which bubbles it up.